Remove LinearAlgebra class. Move methods to appropriate namespaces.
