+++ src/GShark.Test.XUnit/Core/MathTests.cs
@@ -46,5 +46,23 @@ namespace GShark.Test.XUnit.Core
         {
             GSharkMath.IsValidDouble(val).Should().Be(expectedResult);
         }
+
+        [Theory]
+        [InlineData(0, 0, 1)]
+        [InlineData(1, 1, 1)]
+        [InlineData(0, 1, 0)]
+        [InlineData(3, 1, 3)]
+        [InlineData(3, 3, 1)]
+        [InlineData(5, 3, 10)]
+        [InlineData(5, 4, 5)]
+        [InlineData(8, 4, 70)]
+        public void It_Returns_A_Binomial_Coefficient(int n, int k, double resultValue)
+        {
+            // Act
+            double valToCheck = GSharkMath.GetBinomial(n, k);
+
+            // Assert
+            (System.Math.Abs(valToCheck - resultValue) < GSharkMath.Epsilon).Should().BeTrue();
+        }
     }
 }
